Improvements (faster visibility test) Refactor (double[3] to Vector3D, .manager to .data, public to protected, class re-organization) DD is constant now (1 deg = 1000000 dd)

+++ pTolemy3DViewer/branches/refactor-0.1.1/src/org/ptolemy3d/jp2/fast/EntropyDecoder.java
@@ -1511,7 +1511,7 @@ class EntropyDecoder
 
 	private int readByte() {
 		if(pos < count) {
-			return (int)buffer[pos++] & 0xFF;
+			return buffer[pos++] & 0xFF;
 		}
 		else {
 			return -1;
